Added a spot-like capabilities to emissive closure
Emission has two arguments now. Floating angles to define the
spot cone (inner and outer). So we have a penumbra area.
Default bahaviour is a pi/2 cone with not penumbra (a lambert
emitter). The only thing changed in this mode is that now the
normalization factor for that case is 1/pi instead of 1/(2pi)

little fix in smooth decay

Joined the cone emission into a generic one

parameter fixes