New runtime optimization: Track which messages are set by setmessage(), including if a non-constant named message is ever set (could be anything).  Then for getmessage() of a known name, if it couldn't possibly have been set, turn the getmessage into simply an assignment of 0 to the result (not found).
This saves 10% of total render time on our favorite production
benchmark!  Lots of our production shaders check for messages that
trigger special behavior (usually for debugging), but apparently it's
rarely set.  Not only do we save the trouble of doing the getmessage()
in these cases, but often the absence of the message allows us to get
rid of entire "if" clauses of code that won't be used for this shader
instance.

http://codereview.appspot.com/1106041/show
Review by: ckulla