Alter the OSL grammar slightly to allow function prototypes.
Add lots of prototypes to stdosl.h (which is read upon startup
automatically) and remove the corresponding table entries from
typecheck.cpp.  (Not done yet... it will be finished in subsequent
checkins and eventually the whole table may disappear.)

Also make the grammar allow "metadata" about those function
declarations, and use them to communicate other things in the table,
such as [[ int builtin=1 ]] to indicate it's a declaration of a
"builtin" function.

Review by: ckulla
http://codereview.appspot.com/1722049